+++ src/api/contract/registerAndDepositFromStarkTx.js
@@ -1,11 +1,15 @@
 const errorReasons = require('../../lib/dvf/errorReasons')
 const sendToDVFInterface = require('./sendToDVFInterface')
 const { fromQuantizedToBaseUnitsBN } = require('dvf-utils')
+const permitParamsToArgs = require('../../lib/util/permitParamsToArgs')
 
-module.exports = async (dvf, deFiSignature, { starkKey, tokenId, vaultId, amount, tokenAddress, quantum }, options) => {
+module.exports = async (dvf, deFiSignature, { starkKey, tokenId, vaultId, amount, tokenAddress, quantum, permitParams = null }, options) => {
   const ethTokenInfo = dvf.token.getTokenInfoOrThrow('ETH')
+  let action
   if (ethTokenInfo.starkTokenId === tokenId) {
     action = 'registerAndDepositEth'
+  } else if (permitParams) {
+    action = 'registerAndDepositWithPermit'
   } else {
     action = 'registerAndDeposit'
   }
@@ -17,7 +21,7 @@ module.exports = async (dvf, deFiSignature, { starkKey, tokenId, vaultId, amount
     ? [methodArgs, fromQuantizedToBaseUnitsBN(ethTokenInfo, amount).toString()]
     // For other tokens, use amount as is (should be quantised), and add to
     // method args.
-    : [methodArgs.concat(amount.toString(), tokenAddress, quantum)]
+    : [methodArgs.concat(amount.toString(), tokenAddress, quantum).concat(permitParamsToArgs(permitParams))]
 
   return sendToDVFInterface(dvf)(action)(args, value, options)
 }