+++ src/lib/stark/ledger/createFastWithdrawalPayload.js
@@ -81,7 +81,7 @@ module.exports = async (dvf, withdrawalData, path) => {
 
   const tokenContractAddress = token === 'ETH'
     ? address0
-    : tokenInfo.tokenAddress
+    : tokenInfo.tokenAddressPerChain.ETHEREUM
   const quantisedAmount = toQuantizedAmountBN(tokenInfo, amount)
   const baseUnitsAmount = fromQuantizedToBaseUnitsBN(tokenInfo)(quantisedAmount)