+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.5.2 - 2021-06-11
+### Changed
+- Use GET method when getting gas price from API in `getGasPrice`
+
+## 2.5.1 - 2021-06-09
+### Changed
+- Add optional `memo` to transfer API
+
+## 2.5.0 - 2021-06-7
+### Changed
+- Adding optional `permitParams` argument to `depositV2` and `registerAndDeposit` (will try to use signature-based eip-2612 permit instead of `approve` if passed)
+
+## 2.4.0 - 2021-06-4
+### Changed
+- Accomodating changes in configuration tokenRegistry structure (`tokenAddressPerChain`)
+- `dvf.approve` to accept an extra argument `chain` (defaults to `'ETHEREUM'`)
+- accept transaction hash callback option when doing token approval, deposit and register + deposit
+
 ## 2.3.0 - 2021-05-25
 ### Changed
 - `dvf.depositV2` and `dvf.registerAndDeposit` to accept optional `web3Options` parameter for internal use (better gas limit estimates)